+++ src/org/jetbrains/plugins/scala/lang/psi/light/scala/ScLightFunctionDefinition.scala
@@ -10,7 +10,7 @@ import org.jetbrains.plugins.scala.lang.psi.api.expr.{ScAnnotation, ScExpression
 import org.jetbrains.plugins.scala.lang.psi.api.statements.ScFunctionDefinition
 import org.jetbrains.plugins.scala.lang.psi.api.statements.params._
 import org.jetbrains.plugins.scala.lang.psi.types.ScType
-import org.jetbrains.plugins.scala.lang.psi.types.nonvalue.TypeParameter
+import org.jetbrains.plugins.scala.lang.psi.types.api.TypeParameter
 import org.jetbrains.plugins.scala.lang.psi.types.result.{Success, TypeResult}
 
 /**
