+++ src/org/jetbrains/plugins/scala/lang/psi/light/scala/ScLightTypeParamClause.scala
@@ -4,7 +4,7 @@ package lang.psi.light.scala
 import com.intellij.psi.impl.light.LightElement
 import org.jetbrains.plugins.scala.lang.psi.ScalaPsiElement
 import org.jetbrains.plugins.scala.lang.psi.api.statements.params.{ScTypeParam, ScTypeParamClause}
-import org.jetbrains.plugins.scala.lang.psi.types.nonvalue.TypeParameter
+import org.jetbrains.plugins.scala.lang.psi.types.api.TypeParameter
 
 /**
  * @author Alefas
