+++ src/org/jetbrains/plugins/scala/lang/psi/types/Signature.scala
@@ -10,8 +10,8 @@ import org.jetbrains.plugins.scala.extensions._
 import org.jetbrains.plugins.scala.lang.psi.api.statements.params.ScParameters
 import org.jetbrains.plugins.scala.lang.psi.api.statements.{ScFunction, ScTypeAlias, ScTypeAliasDefinition}
 import org.jetbrains.plugins.scala.lang.psi.api.toplevel.ScTypedDefinition
-import org.jetbrains.plugins.scala.lang.psi.types.api.{Any, TypeSystem}
-import org.jetbrains.plugins.scala.lang.psi.types.nonvalue.{TypeParameter, TypeParameterExt}
+import org.jetbrains.plugins.scala.lang.psi.types.api.{Any, TypeParameter, TypeSystem}
+import org.jetbrains.plugins.scala.lang.psi.types.nonvalue.TypeParameterExt
 
 import scala.collection.mutable.ArrayBuffer
