+++ src/org/jetbrains/plugins/scala/lang/psi/types/api/TypeParameterType.scala
@@ -8,6 +8,80 @@ import org.jetbrains.plugins.scala.lang.psi.api.statements.params.ScTypeParam
 import org.jetbrains.plugins.scala.lang.psi.impl.ScalaPsiManager
 import org.jetbrains.plugins.scala.lang.psi.types.{NamedType, ScSubstitutor, ScType, ScUndefinedSubstitutor}
 
+/**
+  * Class representing type parameters in our type system. Can be constructed from psi.
+  * todo: lower and upper types will be reevaluated many times, is it good or bad? Seems bad. What other ways to fix SCL-7216?
+  *
+  * @param lowerType important to be lazy, see SCL-7216
+  * @param upperType important to be lazy, see SCL-7216
+  */
+class TypeParameter(val name: String, val typeParams: Seq[TypeParameter], val lowerType: () => ScType,
+                    val upperType: () => ScType, val ptp: PsiTypeParameter) {
+  def this(ptp: PsiTypeParameter) {
+    this(ptp match {
+      case tp: ScTypeParam => tp.name
+      case _ => ptp.getName
+    }, ptp match {
+      case tp: ScTypeParam => tp.typeParameters.map(new TypeParameter(_))
+      case _ => Seq.empty
+    }, ptp match {
+      case tp: ScTypeParam => () => tp.lowerBound.getOrNothing
+      case _ => () => Nothing //todo: lower type?
+    }, ptp match {
+      case tp: ScTypeParam => () => tp.upperBound.getOrAny
+      case _ => () => Any //todo: upper type?
+    }, ptp)
+  }
+
+  def nameAndId = getPsiElementId(ptp)
+
+  def update(fun: ScType => ScType): TypeParameter = {
+    new TypeParameter(name, typeParams.map(_.update(fun)), {
+      val res = fun(lowerType())
+      () => res
+    }, {
+      val res = fun(upperType())
+      () => res
+    }, ptp)
+  }
+
+  def canEqual(other: Any): Boolean = other.isInstanceOf[TypeParameter]
+
+  override def equals(other: Any): Boolean = other match {
+    case that: TypeParameter =>
+      (that canEqual this) &&
+        name == that.name &&
+        typeParams == that.typeParams &&
+        lowerType() == that.lowerType() &&
+        upperType() == that.upperType() &&
+        ptp == that.ptp
+    case _ => false
+  }
+
+  override def hashCode(): Int = {
+    val state = Seq(name, typeParams, ptp)
+    state.map(_.hashCode()).foldLeft(0)((a, b) => 31 * a + b)
+  }
+}
+
+object TypeParameter {
+  def apply(name: String, typeParams: Seq[TypeParameter], lowerType: () => ScType, upperType: () => ScType,
+            ptp: PsiTypeParameter): TypeParameter = {
+    new TypeParameter(name, typeParams, lowerType, upperType, ptp)
+  }
+
+  def unapply(t: TypeParameter): Option[(String, Seq[TypeParameter], () => ScType, () => ScType, PsiTypeParameter)] = {
+    Some(t.name, t.typeParams, t.lowerType, t.upperType, t.ptp)
+  }
+
+  def fromArray(ptps: Array[PsiTypeParameter]): Array[TypeParameter] = {
+    if (ptps.length == 0) EMPTY_ARRAY
+    else ptps.map(new TypeParameter(_))
+  }
+
+  val EMPTY_ARRAY: Array[TypeParameter] = Array.empty
+}
+
 case class TypeParameterType(name: String, arguments: Seq[TypeParameterType],
                              lower: Suspension[ScType], upper: Suspension[ScType],
                              typeParameter: PsiTypeParameter) extends ValueType with NamedType {
